[UPDATE] New folders organization
