+++ examples/datasets/classifier_dataset.csv
@@ -1,0 +1,82 @@
+h,Temperatura,Velocidade_angular,Pressao_entrada,T_margin
+3,293,10000,1000000,0
+3,293,10000,5500000,0
+3,293,10000,10000000,0
+3,293,15000,1000000,0
+3,293,15000,5500000,0
+3,293,15000,10000000,0
+3,293,20000,1000000,0
+3,293,20000,5500000,0
+3,293,20000,10000000,0
+3,358,10000,1000000,1
+3,358,10000,5500000,0
+3,358,10000,10000000,0
+3,358,15000,1000000,1
+3,358,15000,5500000,0
+3,358,15000,10000000,0
+3,358,20000,1000000,1
+3,358,20000,5500000,0
+3,358,20000,10000000,0
+3,423,10000,1000000,1
+3,423,10000,5500000,1
+3,423,10000,10000000,1
+3,423,15000,1000000,1
+3,423,15000,5500000,1
+3,423,15000,10000000,1
+3,423,20000,1000000,1
+3,423,20000,5500000,1
+3,423,20000,10000000,1
+4,293,10000,1000000,0
+4,293,10000,5500000,0
+4,293,10000,10000000,0
+4,293,15000,1000000,0
+4,293,15000,5500000,0
+4,293,15000,10000000,0
+4,293,20000,1000000,0
+4,293,20000,5500000,0
+4,293,20000,10000000,0
+4,358,10000,1000000,1
+4,358,10000,5500000,0
+4,358,10000,10000000,0
+4,358,15000,1000000,1
+4,358,15000,5500000,0
+4,358,15000,10000000,0
+4,358,20000,1000000,1
+4,358,20000,5500000,0
+4,358,20000,10000000,0
+4,423,10000,1000000,1
+4,423,10000,5500000,1
+4,423,10000,10000000,1
+4,423,15000,1000000,1
+4,423,15000,5500000,1
+4,423,15000,10000000,1
+4,423,20000,1000000,1
+4,423,20000,5500000,1
+4,423,20000,10000000,1
+5,293,10000,1000000,0
+5,293,10000,5500000,0
+5,293,10000,10000000,0
+5,293,15000,1000000,0
+5,293,15000,5500000,0
+5,293,15000,10000000,0
+5,293,20000,1000000,0
+5,293,20000,5500000,0
+5,293,20000,10000000,0
+5,358,10000,1000000,1
+5,358,10000,5500000,0
+5,358,10000,10000000,0
+5,358,15000,1000000,1
+5,358,15000,5500000,0
+5,358,15000,10000000,0
+5,358,20000,1000000,1
+5,358,20000,5500000,0
+5,358,20000,10000000,0
+5,423,10000,1000000,1
+5,423,10000,5500000,1
+5,423,10000,10000000,1
+5,423,15000,1000000,1
+5,423,15000,5500000,1
+5,423,15000,10000000,1
+5,423,20000,1000000,1
+5,423,20000,5500000,1
+5,423,20000,10000000,1
