+++ examples/datasets/regression_dataset.csv
@@ -1,0 +1,82 @@
+h,Temperatura,Velocidade_angular,Pressao_entrada,T_margin
+3,293,10000,1000000,-37.551749
+3,293,10000,5500000,-65.719359
+3,293,10000,10000000,-65.448138
+3,293,15000,1000000,-40.893463
+3,293,15000,5500000,-64.57492
+3,293,15000,10000000,-68.152387
+3,293,20000,1000000,-42.196224
+3,293,20000,5500000,-68.792711
+3,293,20000,10000000,-73.693987
+3,358,10000,1000000,25.738684
+3,358,10000,5500000,4.533542
+3,358,10000,10000000,0.631433
+3,358,15000,1000000,23.972517
+3,358,15000,5500000,1.098388
+3,358,15000,10000000,0.514589
+3,358,20000,1000000,22.314472
+3,358,20000,5500000,-2.98075
+3,358,20000,10000000,-0.953247
+3,423,10000,1000000,89.538887
+3,423,10000,5500000,69.728928
+3,423,10000,10000000,69.050971
+3,423,15000,1000000,87.746133
+3,423,15000,5500000,66.034165
+3,423,15000,10000000,62.785178
+3,423,20000,1000000,86.329235
+3,423,20000,5500000,62.115823
+3,423,20000,10000000,66.62483
+4,293,10000,1000000,-36.532524
+4,293,10000,5500000,-60.370185
+4,293,10000,10000000,-64.927413
+4,293,15000,1000000,-38.951996
+4,293,15000,5500000,-63.675058
+4,293,15000,10000000,-79.390178
+4,293,20000,1000000,-41.228791
+4,293,20000,5500000,-67.354983
+4,293,20000,10000000,-73.279079
+4,358,10000,1000000,27.912347
+4,358,10000,5500000,4.751035
+4,358,10000,10000000,-2.16505
+4,358,15000,1000000,25.802301
+4,358,15000,5500000,1.782496
+4,358,15000,10000000,1.423944
+4,358,20000,1000000,22.590301
+4,358,20000,5500000,-2.13356
+4,358,20000,10000000,-0.445033
+4,423,10000,1000000,91.960553
+4,423,10000,5500000,69.948987
+4,423,10000,10000000,73.015295
+4,423,15000,1000000,90.414615
+4,423,15000,5500000,66.839969
+4,423,15000,10000000,70.674416
+4,423,20000,1000000,87.43037
+4,423,20000,5500000,62.720855
+4,423,20000,10000000,67.45217
+5,293,10000,1000000,-36.456564
+5,293,10000,5500000,-60.411013
+5,293,10000,10000000,-79.954509
+5,293,15000,1000000,-38.691902
+5,293,15000,5500000,-63.767393
+5,293,15000,10000000,-66.98147
+5,293,20000,1000000,-41.627576
+5,293,20000,5500000,-66.583097
+5,293,20000,10000000,-71.120044
+5,358,10000,1000000,28.443254
+5,358,10000,5500000,4.98845
+5,358,10000,10000000,-2.128808
+5,358,15000,1000000,26.069309
+5,358,15000,5500000,1.903157
+5,358,15000,10000000,1.329993
+5,358,20000,1000000,22.991674
+5,358,20000,5500000,-2.913819
+5,358,20000,10000000,1.898159
+5,423,10000,1000000,93.731372
+5,423,10000,5500000,69.959863
+5,423,10000,10000000,69.711589
+5,423,15000,1000000,90.919624
+5,423,15000,5500000,66.937906
+5,423,15000,10000000,70.573387
+5,423,20000,1000000,87.703816
+5,423,20000,5500000,63.107396
+5,423,20000,10000000,68.549709
